Add ORDER BY support and instruction limit safety
- Implement sorter opcodes (SorterInsert, SorterSort, SorterNext, SorterData)
- Add DecodeRecord opcode for extracting columns from serialized records
- Add instruction counter with 100K limit to prevent infinite loops
- Fix flaky PRNG test to use local state instead of shared global
- Various executor and VDBE improvements for test compatibility

+++ src/os/unix.rs
@@ -181,7 +181,7 @@ impl Vfs for UnixVfs {
             flags,
             lock_type: UnsafeCell::new(LockType::None),
             shm: UnsafeCell::new(None),
-            dev_ino: (stat.st_dev as u64, stat.st_ino as u64),
+            dev_ino: (stat.st_dev, stat.st_ino),
             sector_size: 4096,
             device_chars: self.detect_device_characteristics(fd),
             delete_on_close,
@@ -500,7 +500,7 @@ impl VfsFile for UnixFile {
         if rc != 0 {
             return Err(UnixVfs::error_from_errno());
         }
-        Ok(stat.st_size as i64)
+        Ok(stat.st_size)
     }
 
     fn lock(&self, lock_type: LockType) -> Result<()> {
